move font loading to root
this allows font loading to be separated from other reloading things in a child route
